+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # orchid-core
 
+## 0.1.3
+
+### Patch Changes
+
+- Support database CHECK validation
+
 ## 0.1.2
 
 ### Patch Changes