+++ packages/qb/myqb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # myqb
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes